Added mocha tests, full package.json, refactor
