+++ package.json
@@ -1,9 +1,32 @@
 {
+  "name": "kelp",
+  "version": "0.0.1",
+  "description": "An easy to use, customizable ExpressJS web server.",
+  "main": "index.js",
+  "scripts": {
+    "test": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/znci/kelp.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/znci/kelp/issues"
+  },
+  "homepage": "https://github.com/znci/kelp#readme",
   "dependencies": {
     "body-parser": "^1.20.2",
     "chalk": "^4.1.2",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "pug": "^3.0.2"
+  },
+  "devDependencies": {
+    "chai": "^4.3.7",
+    "mocha": "^10.2.0",
+    "nodemon": "^2.0.12",
+	"node-fetch": "^2.6.9"
   }
 }
