+++ public/index.html
@@ -16,7 +16,7 @@
 		}
 		/* weight 900 h1 */
 		h1 {
-			font-weight: 900;
+			font-weight: 700;
 			font-size: 2.5em;
 			margin: 0.67em 0;
 		}
